Move friend notification from App to Header
