+++ src/App.tsx
@@ -1,17 +1,9 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { Outlet } from "react-router-dom";
 import { createGlobalStyle } from "styled-components";
-import {
-  doc,
-  query,
-  collection,
-  where,
-  onSnapshot,
-  getFirestore,
-} from "firebase/firestore";
 import { auth } from "../src/utils/firebase";
 import { onAuthStateChanged } from "firebase/auth";
-import { useSelector, useDispatch } from "react-redux";
+import { useDispatch } from "react-redux";
 import { setUserData } from "../src/actions/index";
 import firebaseapi from "../src/utils/firebaseapi";
 
@@ -19,7 +11,6 @@ import Header from "./components/Header";
 import Footer from "./components/Footer";
 import Slider from "./components/Slider";
 import ScrollToTop from "./components/ScrollToTop";
-import Notification from "./components/modal/Notification";
 
 const GlobalStyle = createGlobalStyle`
   * {
@@ -44,11 +35,7 @@ a {
 `;
 
 function App() {
-  const db = getFirestore();
-  const [showNotification, setShowNotification] = useState(false);
   const dispatch = useDispatch();
-  const [getInvitationList, setGetInvitationList] = useState<any>();
-  const [arrayLength, setArrayLength] = useState(0);
 
   useEffect(() => {
     onAuthStateChanged(auth, (user) => {
@@ -63,34 +50,10 @@ function App() {
                 result["main_photo"]
               )
             );
-            // see friend requests difference
-            getFriend(result["user_id"]);
-            //
-            // see chatroom update
-            //
           }
         });
       }
     });
-  }, []);
-
-  const getFriend = (id: string) => {
-    onSnapshot(doc(collection(db, "Users"), id), (doc) => {
-      if (doc.exists()) {
-        setGetInvitationList(doc.data().friend_request);
-        // console.log(doc.data().friend_request);
-        // To compare the friend request
-        setArrayLength(doc.data().friend_request.length);
-        setShowNotification(false);
-        if (doc.data().friend_request.length > arrayLength) {
-          setShowNotification(true);
-        }
-      }
-    });
-  };
-
-  useEffect(() => {
-    // console.log("new PR!");
   }, []);
 
   return (
@@ -98,7 +61,6 @@ function App() {
       <GlobalStyle />
       <Header />
       <Outlet />
-      {showNotification && <Notification />}
       <Slider />
       <ScrollToTop />
       <Footer />
