+++ src/components/Header.tsx
@@ -4,12 +4,22 @@ import logo from "./logo.png";
 import { useLocation, Link, useNavigate } from "react-router-dom";
 import { useSelector, useDispatch } from "react-redux";
 import firebaseapi from "../../src/utils/firebaseapi";
-import { DocumentData } from "firebase/firestore";
+import {
+  doc,
+  query,
+  collection,
+  where,
+  onSnapshot,
+  getFirestore,
+  DocumentData,
+} from "firebase/firestore";
 import useOnclickOutside from "react-cool-onclickoutside";
 import { ReactComponent as GitHub } from "./github.svg";
 import { ReactComponent as Member } from "./member.svg";
 import { Tours, stepType } from "./Tours";
 import SearchResults from "./SearchResults";
+
+import Notification from "../components/modal/Notification";
 
 const Wrapper = styled.div`
   display: flex;
@@ -177,12 +187,15 @@ const ClostBtn = styled.button`
 `;
 
 const Header = () => {
+  const db = getFirestore();
+  const [showNotification, setShowNotification] = useState(false);
+  const [getInvitationList, setGetInvitationList] = useState<any>();
+  const [arrayLength, setArrayLength] = useState(0);
   const currentPage = useLocation();
   const [page, setPage] = useState<any>("");
   let navigate = useNavigate();
   const dispatch = useDispatch();
   const userInfo = useSelector((state) => state) as any;
-  const [alreadyLogged, setAlreadyLogged] = useState(false);
 
   const [searchResults, setSearchRsults] = useState<DocumentData>();
 
@@ -193,10 +206,28 @@ const Header = () => {
   };
 
   useEffect(() => {
+    const userId = userInfo.user.user_id;
+    console.log(userId);
+    getFriend(userId);
     // console.log(currentPage);
     // console.log(currentPage.pathname);
     setPage(currentPage.pathname);
   }, [currentPage]);
+
+  const getFriend = (id: string) => {
+    onSnapshot(doc(collection(db, "Users"), id), (doc) => {
+      if (doc.exists()) {
+        setGetInvitationList(doc.data().friend_request);
+        // console.log(doc.data().friend_request);
+        // To compare the friend request
+        setArrayLength(doc.data().friend_request.length);
+        setShowNotification(false);
+        if (doc.data().friend_request.length > arrayLength) {
+          setShowNotification(true);
+        }
+      }
+    });
+  };
 
   const memberHandler = () => {
     if (userInfo.user.user_id) {
@@ -297,6 +328,7 @@ const Header = () => {
               </Category>
             </LeftContainer>
             <RightContainer>
+              {showNotification && <Notification />}
               <Category as="div" id="docs">
                 <Tours stepType={stepType} page={page} />
               </Category>
@@ -306,9 +338,6 @@ const Header = () => {
               <Category as="div" onClick={memberHandler}>
                 <Member />
               </Category>
-              {/* <Category as="div" onClick={signout}>
-                <Logout />
-              </Category> */}
             </RightContainer>
           </CategoryContainer>
         </CategoryLinks>
